improvement: added supplementary data capability
In order to allow for additional metadata values to be stored
for various file types, added supplementary table capability to
the tika metadata class.

This allows for supplementary tables to be defined in the
metadataextractor_tika subplugin, so that additional information
can be stored based on the mimetype of the resource as parsed by
tika.